Adds Maze DTO constructor and tests
Adds a constructor to the Maze class that accepts a MazeDto,
allowing the creation of a Maze object from a DTO.

Also, adds unit tests for the Maze class, including tests for
the DTO constructor and the ToDto method.

Removes redundant code in MazeContainer by utilizing the new Maze constructor.
Adds several test classes for factories and handlers.